+++ pages/content-types/index.js
@@ -1,4 +1,5 @@
 import React, { useContext, useEffect, useState } from 'react'
+import Head from 'next/head';
 import NextLink from 'next/link';
 
 // APIs
@@ -170,6 +171,9 @@ function ContentTypes() {
 
   return (
     <>
+      <Head>
+        <title>Content Types</title>
+      </Head>
       <PageTitleWrapper>
         <Grid
           container