+++ CHANGELOG.md
@@ -1,6 +1,11 @@
-### Changelog
+# Changelog
 
 All notable changes to this project will be documented in this file
 
-## 30 March 2023
+## March 31, 2023
+- Update page title to reflect the name of the page
+- Added GETTING-STARTED.md into the application at /getting-started
+- Added CONTRIBUTING.md into the application at /contributing
+
+## March 30, 2023
 - Add release notes from CHANGELOG.md into the application at /release-notes